Modified client

+++ client/src/components/commentCard.js
@@ -46,41 +46,54 @@ class CommentCard extends Component{
         };
     }
     onDeleteClick(id){
-    const datadel = {
-        title:this.state.comment.title,
-        content:this.state.comment.content,
-        author:this.state.comment.author
-    };
-    console.dir(datadel);
-    axios.delete(`http://localhost:8082/api/posts/${this.state.comment.post}/comments/${id}`
-    ,{headers: authHeader(),data:datadel,params:{"secret_token":AuthService.getCurrentUser()}}
-    )
-    .then(res =>{
-        this.props.history.push(`/show-post/${this.state.comment.post}/show-comments`);
-    })
-    .catch(err =>{
-        console.log("Error from commentcard_deleteClick");
-    })
+    if(!AuthService.getUser()){
+        alert("Please login to delete comment");
+    }
+    else{
+        const datadel = {
+            title:this.state.comment.title,
+            content:this.state.comment.content,
+            author:this.state.comment.author
+        };
+        // console.dir(datadel);
+        axios.delete(`http://localhost:8082/api/posts/${this.state.comment.post}/comments/${id}`
+        ,{headers: authHeader(),data:datadel,params:{"secret_token":AuthService.getCurrentUser()}}
+        )
+        .then(res =>{
+            this.props.history.push(`/show-post/${this.state.comment.post}/`);
+        })
+        .catch(err =>{
+            console.log("Error from commentcard_deleteClick");
+            console.log(err);
+        })
+    }
+
 };
   render(){
       return(
-                  <div className="card-container">
+            <div className="card-comment card-container">
             <div className="desc">
                 <h3>{this.state.comment.title}</h3>
                 <h5>{this.state.comment.author}</h5>
                 <p>{this.state.comment.content}</p>
+                <p>{this.state.comment.createdAt.slice(0,10)}</p>
             </div>
             <div className="row">
-                <div className="col-md-6">
+                <div className="col-md-1">
                     <Link to={`/show-post/${this.state.comment.post}/edit-comment/${this.state.comment._id}`}
-                    className="btn btn-outline-info btn-lg btn-block">Edit comment</Link>
+                    className="btn btn-outline-warning btn-lg btn-block">
+                        <i className="fa fa-pencil-square-o" aria-hidden="true">Edit</i>
+                    </Link>
                 </div>
-                <div className="col-md-6">
+                <div className="col-md-1">
                     <button type="button" className="btn btn-outline-danger btn-lg btn-block" 
-                    onClick={this.onDeleteClick.bind(this,this.state.comment._id)}>Delete comment</button>
+                    onClick={this.onDeleteClick.bind(this,this.state.comment._id)}>
+                        <i className="fa fa-trash" aria-hidden="true"></i>
+                    </button>
                 </div>
             </div>
         </div>
+       
       )
   }
 }
